+++ dubbo-dependencies-bom/pom.xml
@@ -133,6 +133,8 @@
         <micrometer.version>1.7.4</micrometer.version>
         <t_digest.version>3.3</t_digest.version>
         <prometheus_client.version>0.10.0</prometheus_client.version>
+        <reactive.version>1.0.4</reactive.version>
+        <reactor.version>3.4.19</reactor.version>
 
         <rs_api_version>2.0</rs_api_version>
         <resteasy_version>3.0.19.Final</resteasy_version>
@@ -801,6 +803,17 @@
                 <groupId>io.prometheus</groupId>
                 <artifactId>simpleclient_pushgateway</artifactId>
                 <version>${prometheus_client.version}</version>
+            </dependency>
+            <!-- reactive related dependencies -->
+            <dependency>
+                <groupId>org.reactivestreams</groupId>
+                <artifactId>reactive-streams</artifactId>
+                <version>${reactive.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.projectreactor</groupId>
+                <artifactId>reactor-core</artifactId>
+                <version>${reactor.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
