+++ dubbo-distribution/dubbo-bom/pom.xml
@@ -173,6 +173,11 @@
                 <artifactId>dubbo-container-spring</artifactId>
                 <version>${project.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.apache.dubbo</groupId>
+                <artifactId>dubbo-reactive</artifactId>
+                <version>${project.version}</version>
+            </dependency>
 
             <!-- dubbo plugin -->
             <dependency>
