+++ dubbo-plugin/dubbo-reactive/src/main/java/org/apache/dubbo/reactive/ServerTripleReactorSubscriber.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.rpc.protocol.tri.reactive;
+package org.apache.dubbo.reactive;
 
 import org.apache.dubbo.rpc.CancellationContext;
 import org.apache.dubbo.rpc.protocol.tri.CancelableStreamObserver;
@@ -24,7 +24,7 @@ import org.apache.dubbo.rpc.protocol.tri.observer.CallStreamObserver;
 /**
  * The Subscriber in server to passing the data produced by user publisher to responseStream.
  */
-public class ServerTripleReactorSubscriber<T> extends AbstractTripleReactorSubscriber<T>{
+public class ServerTripleReactorSubscriber<T> extends AbstractTripleReactorSubscriber<T> {
 
     @Override
     public void subscribe(CallStreamObserver<T> downstream) {
