+++ dubbo-plugin/dubbo-reactive/src/main/java/org/apache/dubbo/reactive/handler/ManyToManyMethodHandler.java
@@ -15,35 +15,33 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.rpc.protocol.tri.reactive.handler;
+package org.apache.dubbo.reactive.handler;
 
 import org.apache.dubbo.common.stream.StreamObserver;
-import org.apache.dubbo.rpc.protocol.tri.reactive.calls.ReactorServerCalls;
-import org.apache.dubbo.rpc.stub.FutureToObserverAdaptor;
+import org.apache.dubbo.reactive.calls.ReactorServerCalls;
+import org.apache.dubbo.rpc.protocol.tri.observer.CallStreamObserver;
 import org.apache.dubbo.rpc.stub.StubMethodHandler;
-import reactor.core.publisher.Mono;
+import reactor.core.publisher.Flux;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 
 /**
- * The handler of OneToOne() method for stub invocation.
+ * The handler of ManyToMany() method for stub invocation.
  */
-public class OneToOneMethodHandler<T, R> implements StubMethodHandler<T, R> {
+public class ManyToManyMethodHandler<T, R> implements StubMethodHandler<T, R> {
 
-    private final Function<Mono<T>, Mono<R>> func;
+    private final Function<Flux<T>, Flux<R>> func;
 
-    public OneToOneMethodHandler(Function<Mono<T>, Mono<R>> func) {
+    public ManyToManyMethodHandler(Function<Flux<T>, Flux<R>> func) {
         this.func = func;
     }
 
     @SuppressWarnings("unchecked")
     @Override
-    public CompletableFuture<R> invoke(Object[] arguments) {
-        T request = (T) arguments[0];
-        CompletableFuture<R> future = new CompletableFuture<>();
-        StreamObserver<R> responseObserver = new FutureToObserverAdaptor<>(future);
-        ReactorServerCalls.oneToOne(request, responseObserver, func);
-        return future;
+    public CompletableFuture<StreamObserver<T>> invoke(Object[] arguments) {
+        CallStreamObserver<R> responseObserver = (CallStreamObserver<R>) arguments[0];
+        StreamObserver<T> requestObserver = ReactorServerCalls.manyToMany(responseObserver, func);
+        return CompletableFuture.completedFuture(requestObserver);
     }
 }
