+++ dubbo-plugin/dubbo-reactive/src/main/java/org/apache/dubbo/reactive/handler/ManyToOneMethodHandler.java
@@ -15,25 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.rpc.protocol.tri.reactive.handler;
+package org.apache.dubbo.reactive.handler;
 
 import org.apache.dubbo.common.stream.StreamObserver;
+import org.apache.dubbo.reactive.calls.ReactorServerCalls;
 import org.apache.dubbo.rpc.protocol.tri.observer.CallStreamObserver;
-import org.apache.dubbo.rpc.protocol.tri.reactive.calls.ReactorServerCalls;
 import org.apache.dubbo.rpc.stub.StubMethodHandler;
 import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 
 /**
- * The handler of ManyToMany() method for stub invocation.
+ * The handler of ManyToOne() method for stub invocation.
  */
-public class ManyToManyMethodHandler<T, R> implements StubMethodHandler<T, R> {
+public class ManyToOneMethodHandler<T, R> implements StubMethodHandler<T, R> {
 
-    private final Function<Flux<T>, Flux<R>> func;
+    private final Function<Flux<T>, Mono<R>> func;
 
-    public ManyToManyMethodHandler(Function<Flux<T>, Flux<R>> func) {
+    public ManyToOneMethodHandler(Function<Flux<T>, Mono<R>> func) {
         this.func = func;
     }
 
@@ -41,7 +42,7 @@ public class ManyToManyMethodHandler<T, R> implements StubMethodHandler<T, R> {
     @Override
     public CompletableFuture<StreamObserver<T>> invoke(Object[] arguments) {
         CallStreamObserver<R> responseObserver = (CallStreamObserver<R>) arguments[0];
-        StreamObserver<T> requestObserver = ReactorServerCalls.manyToMany(responseObserver, func);
+        StreamObserver<T> requestObserver = ReactorServerCalls.manyToOne(responseObserver, func);
         return CompletableFuture.completedFuture(requestObserver);
     }
 }
