+++ dubbo-plugin/dubbo-reactive/src/main/java/org/apache/dubbo/reactive/handler/OneToOneMethodHandler.java
@@ -15,34 +15,35 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.rpc.protocol.tri.reactive.handler;
+package org.apache.dubbo.reactive.handler;
 
 import org.apache.dubbo.common.stream.StreamObserver;
-import org.apache.dubbo.rpc.protocol.tri.reactive.calls.ReactorServerCalls;
+import org.apache.dubbo.reactive.calls.ReactorServerCalls;
+import org.apache.dubbo.rpc.stub.FutureToObserverAdaptor;
 import org.apache.dubbo.rpc.stub.StubMethodHandler;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 
 /**
- * The handler of OneToMany() method for stub invocation.
+ * The handler of OneToOne() method for stub invocation.
  */
-public class OneToManyMethodHandler<T, R> implements StubMethodHandler<T, R> {
+public class OneToOneMethodHandler<T, R> implements StubMethodHandler<T, R> {
 
-    private final Function<Mono<T>, Flux<R>> func;
+    private final Function<Mono<T>, Mono<R>> func;
 
-    public OneToManyMethodHandler(Function<Mono<T>, Flux<R>> func) {
+    public OneToOneMethodHandler(Function<Mono<T>, Mono<R>> func) {
         this.func = func;
     }
 
     @SuppressWarnings("unchecked")
     @Override
-    public CompletableFuture<?> invoke(Object[] arguments) {
+    public CompletableFuture<R> invoke(Object[] arguments) {
         T request = (T) arguments[0];
-        StreamObserver<R> responseObserver = (StreamObserver<R>) arguments[1];
-        ReactorServerCalls.oneToMany(request, responseObserver, func);
-        return CompletableFuture.completedFuture(null);
+        CompletableFuture<R> future = new CompletableFuture<>();
+        StreamObserver<R> responseObserver = new FutureToObserverAdaptor<>(future);
+        ReactorServerCalls.oneToOne(request, responseObserver, func);
+        return future;
     }
 }
